django/requirements.txt
Django==3.2.4
psycopg2-binary==2.8.5
djangorestframework==3.12.4
requests==2.25.1
pydantic==1.8.2
pandas==1.2.4
celery==5.1.0
redis==3.5.3
drf-yasg==1.20.0
